Run job using multiprocessing.
This allows us to capture the state of the process in a much nicer
fashion.  Earlier, if the command was executed and completed with an
error but if the automation had been stopped, there was no way to tell
if it was done or had an error.  Now, this is done in a better way.
Also improved the test coverage a little.