The Problem now creates the tasks.
This allows a user to easily customize the tasks to actually run instead
of the current circuitous approach with passing the task_cls all over
the place.  We also retain the get_commands method as it makes it easy
to just spell out a bunch of commands and automatically generate the
tasks.  This can also be customized with the Problem's task_cls class
attribute.