Much better testing for cluster manager.
Test cluster management without mocking on Python 3.x at least.  This
still does not connect to a remote host but does test the bootstrapping,
update, and actual task execution on the remote worker.  Also fixes some
issues where I was inadvertently assuming that the root directory was
not what it was.  This should also increase test coverage a good bit.