BUG: Incorrect calculation of free cores.
When calculating the actual free cores, we were looking at the number of
free cores and subtracting the number of cores used which essentially
double counts the cores that are used since free will be lower when the
cores are being used.  We now subtract the number of used cores from the
total available to decide if we can run the job.