ENH: Allow problems to depend on other problems.
The inter-problem dependency is created by changing the `get_requires`
method of the Problem subclass.  The requires can be either another
task, or a Problem instance or class, or a combination of these.

Also support the case where multiple tasks depend on one task.  In this
case, only one instance of the task is added even if multiple instances
of that task are created.  This is all done based on the "outputs" of a
task.  Two tasks with the same `output()` are treated as the same --
this is consistent with the automan design that each
command/run/simulation goes in its own directory.