Pass the force argument to the SolveProblem task.
This is useful when you have a problem that depends on others.
Currently if you force this problem, then it will not force the other
tasks it depends on.  This means that if that problem needs to be rerun
you have to manually ask for it to be run with `-f`.  This change avoids
all that and passes along the force argument to any dependent problems,