BUG: Fix more subtle errors.
When a problem's requirements did not execute correctly, then the
SolveProblem task was treated as running because the check_status
function would treat even tasks that hadn't started as if they were
running and at the end we would end up waiting for the apparently
running task.  This is now fixed.  The completion logic for cases where
a match condition is set is also fixed.  Also, the complete method for a
command task will always raise an error even when called multiple times
in the same execution.  This is a lot more consistent.