BUG: Issue with remote command task's complete.
This is a subtle issue with command tasks executing on a remote host.
Basically if you have remote hosts, the scheduler would end up waiting
on some tasks which were really already completed.  This was because
those tasks were returning None on their complete method and not True.
This is now tested and fixed.